Moved all lore related methods into a separate class.
